pyproject.toml
[tool.poetry]
name = "feathery"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
pathlib = "^1.0.1"
pdf2image = "^1.17.0"
python-dotenv = "^1.0.1"
openai = "1.1.1"
requests = "^2.32.3"
instructor = "^1.3.4"
fitz = "^0.0.1.dev2"
pymupdf = "^1.24.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
